+++ tools.go
@@ -6,4 +6,5 @@ package tools
 import (
 	_ "github.com/golang/mock/mockgen"
 	_ "sigs.k8s.io/kind"
+	_ "golang.org/x/tools/cmd/godoc"
 )